pyproject.toml
[project]
name = "prompt-similarity"
version = "0.3.0"
description = "Semantic similarity search and duplicate detection for prompt libraries"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "openai>=1.50.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
ui = [
    "streamlit>=1.38.0",
    "plotly>=5.22.0",
]
eval = [
    "requests>=2.32.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
prompt-similarity = "prompt_similarity.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prompt_similarity"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
